feat : 구독 요청 시, 해당 유저의 알림 개수 로직 구현

+++ src/main/java/com/sosim/server/notification/dto/response/NotificationCountResponse.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 @Builder
 public class NotificationCountResponse {
 
-    private int count;
+    private long count;
 
-    public static NotificationCountResponse toDto(int count) {
+    public static NotificationCountResponse toDto(long count) {
         return NotificationCountResponse.builder()
                 .count(count)
                 .build();
